Update 2013-08-01-role.md
A real eye-opener... thanks Ratnesh for such an awesomest article...


I am not comfortable about the placements of the 22nd and 24th segments, should they be at a different place????  "One thing stands out clearly. We were forced to do so many things against our very nature. Eat what we did not like the taste of, study books that were horror, meet people whom we wanted to run away from, et al.

The moral of the story is that as parents we constantly swim upstream (and achieve so less)."...